requirements.txt
Flask==2.3.3
snowflake-connector-python==3.4.0
bcrypt==4.0.1
requests==2.31.0
sqlparse==0.4.4
python-dotenv==1.0.0
gunicorn==21.2.0
